power: supply: qti_battery_charger: Add support for FLASH_ACTIVE property
Add support for FLASH_ACTIVE property, which is used to notify
the charger when flash or torch are in use.

While propagating to 6.6 kernel, also add the USB_SCOPE,
USB_CONNECTOR_TYPE properties in the usb_property_id enum
to keep it consistent with 5.10 kernel and match with the ones
used on ADSP firmware side.

Change-Id: I429f8b5fb15ff206cf75b536a03c463d8668a652
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>
Signed-off-by: Shilpa Suresh <quic_c_sbsure@quicinc.com>