remoteproc: q6v5: Move the crash work to unbounded workqueue
Currently, if recovery is disabled, crash handler work is bounded work
which is a difference compared to what is being done if recovery is
enabled and it is possible some work is already being running and not
proceeding on the cpu where remote processor crash handler is queued.
Queueing the work on the unbounded queue gives a great chance to the
work to get handled on other cpu.

Change-Id: I1fd90541a244ef377c8421ebfc4188386492b416
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>