gunyah: Add support for older versions of gunyah
This patch adds support for older versions of gunyah hypervisor.

A dummy VM image auth RM call is made during the gunyah_loader probe,
which is expected to always fail but with different error values. On
older versions on Gunyah, it fails with -EOPNOTSUPP.
A global variable is set to true, if return value matches -EOPNOTSUPP.
This variable is then used at various places to load and authenticate VM
images in ways suitable for older firmware and opt out of RPCs which are
not supported by older gunyah.

Change-Id: I9fd150733ec9b214fb441ee024219a0cbf633075
Signed-off-by: Hrishabh Rajput <quic_hrishabh@quicinc.com>