sched: walt: modify yield guard threshold for pipeline case
Currently, a task must exceed a specific high yield threshold
to be classified as a frequently yielding task. However, it has
been noted that during pipeline use cases, tasks often fail to
meet this high threshold despite frequently yielding.

To address this, the logic for detecting frequent yielders is
updated to incorporate two distinct thresholds for pipeline
and non-pipeline scenarios. This adjustment will ensure that
tasks under pipeline conditions are recognized as frequent
yielders at a lower threshold, while in standard scenarios,
only tasks with a very high frequency of yielding are identified.

Change-Id: I075ee39d01e31622d6b9fa2520df7b2ec4b76fbe
Signed-off-by: Yufan Zhu <quic_yufazhu@quicinc.com>
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>