usb: phy: m31: Add cable connection check in set_power
When we perform composition switch or adb root, USB_ICL will endup
voting 0mA. When composition switch is happening it invokes composite
disconnect which internally invokes vbus_draw with 0mA and it result
in system shutdown when device is running with low battery.

To prevent this add a check not to vote for 0mA when cable is connected.

The expectation should be to notify PHY disconnect before disabling the
vbus override. When disabling vbus override, the disconnect IRQ is
received which will vote for 0mA. In that case, make sure PHY driver
knows cable has been disconnected.

Change-Id: I539f5a75dd08e28cbbd1950c3867686e259076f9
Signed-off-by: Pratham Pratap <quic_ppratap@quicinc.com>