phy: msm: snps: eusb2: Update the eusb2 init sequence
Currently the driver is following only one init sequence for all chipsets
which seem to be wrong as the latest hpg revision describes additional
changes in the sequence which is missing.

This change add the missing register writes into the eusb2 init sequence
to make sure of the proper functionality.

Description:

USB_PHY_CFG0: setting to 0x00 at the end of init sequence.

Change-Id: I171d257548ad5b3065250807678a77924fb7570c
Signed-off-by: Uttkarsh Aggarwal <quic_uaggarwa@quicinc.com>