ANDROID: build: Add output_opts.txt file to output
Bazel builds use a cache dir. Because it is unset,
it is inconsistent between the kernel build and
building a DDK module. We can reconcile the
differences by adding `--output_user_root=${bazel-cache-dir}`
to the DDK module build. Create a new file,
output_opts.txt, which contains this flag
and can be used by the DDK build_module.sh.

Change-Id: I68ae0e9a6723988e910465f2a9c59867b2eb2ac9
Signed-off-by: Yuanjie Yang <quic_yuanjiey@quicinc.com>