sched/walt: Fix pipeline misfit behavior under prime worthy load
At the introduction of commit ef90f13d7ea5 ("sched/walt: Disable
misfits for pipeline tasks"), pipeline misfits were disabled.
As a consequence of this, is_prime_worthy is now broken, as
pipeline tasks will never be treated as tasks that can fit on primes.

In order to fix this, ensure that any task with a prime pipeline CPU is
considered as a misfit task on non-prime CPUs.

In addition, ensure that if the number of pipeline tasks is 2 or less,
tasks are given the opportunity to become misfit.

Furthermore, ensure that this behavior is agnostic to special pipeline
task - in other words, if special pipeline task is involved and nr=2,
and one of the two tasks becomes prime worthy, ensure the prime worthy
task is still moved to the prime CPU.

Change-Id: I583c2bdabd1cc9a75cdeb863dc0188a96b115a9f
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>