remoteproc: q6v5: do msleep and panic under rproc lock
msleep without rproc lock of a crash subsystem is not a
safe option to do for a case if recovery is disabled for
that rproc as there is chance there is a parallel request
is to make that rproc up and come and eventually crash as
the regions are not unlocked when crash happens for rproc.

Let's not allow the other thread to take over by using
msleep and panic inside rproc lock.

Change-Id: I4cad1d7481dc9bd9d34bedc02d7e2377681e9603
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>