build: msm_kernel_le: Add signing/verity keys to output
Signing keys and certificates are required to sign kernel modules
and dm-verity hash, but these files are not generated in the current
msm_kernel_le build.

Add a Bazel rule to generate them and add to the output of the le
distributions. Also add the scripts and pre-generated key required
to use them.

Change-Id: I686c90ece024f857edec8b81baa5eecada9cb64c
Signed-off-by: Khaja Hussain Shaik Khaji <quic_kshaikkh@quicinc.com>