power: reset: qcom_dload: Fix kasan reported OOB error
Add fix to change enable_dump to int to fetch the value of the
enable_dump variable into the kernel parameter. Referencing boolean
as an integer was causing the out of bound kasan error. Also add  clean-up
to remove the unneeded function which is exported to get the dump_mode
variable as it is not used anywhere and the sysfs interface already exposes
the same.

Change-Id: I1e42f96e4f7383602b4bc41ff5cf684fdc6b3834
Signed-off-by: Souradeep Chowdhury <quic_schowdhu@quicinc.com>