usb: phy: m31: Address higher current consumption in m31 eusb phy
Currently the driver is setting the zeroth bit of USB_PHY_CFG1
responsible for controlling PLL clock sources. This causes the
phy to draw more current in host mode bus suspend showing a higher
consumption. However this had been addressed in the Phy
documentation is recommended not to set this bit.

Addressing the documentation recommended setting in order to fix
the current consumption issue.

Change-Id: I1d2e23d98b1fe7437b83f520f5b65bd9ea1b4430
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>