usb: phy: m31: Removing the ref_clk_enable flag from m31 phy
Currently, the ref_clk_enable flag is used to control the ref_clk
during set_suspend callback. This allows the clock to be on in
suspend to detect resume signalling in bus suspend cases. However
this result in a higher power consumption which isn't favoured.

Fix this by removing the snippet for keeping the clock enabled
as well as cleaning the code.

Change-Id: If7e1b9aa8f677b65f0bf853ed86dfa0dd23fe2f2
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>