usb: phy: Turn on PHY power when EUD is enabled in XBL
When the device is booted with EUD enabled in XBL, dwc3 probe fails.
This results in USB functionality and EUD not working because the PHY
is not re-initialized while EUD is active. In this change if suspend bit
is not "1" then to keep EUD session alive LDOs should be "on".

Change-Id: Ie82bc64aa4bf42a71b903630cd772f9b558e8b15
Signed-off-by: Uttkarsh Aggarwal <quic_uaggarwa@quicinc.com>