dmaengine: gpi: Enhanced Timeout Handling with Register State Verification
For gpii event or channel command if we rely on interrupts, sometimes
due to CPU busy, IRQ is served with delay, which is causing cmd
timeout which negatively impacts the reliability of command execution.

To address this issue, we have introduced a mechanism to directly read
the relevant hardware register when a command completion timeout occurs.
This approach allows us to bypass the dependency on interrupts and
verify the command status directly from the hardware.

Specifically, when the hardware completes a command, it sets the
CH_ctrl_irq bit in the control register. By reading this bit, we can
confirm that the hardware has successfully processed the command.
If the CH_ctrl_irq bit is set, we return success, indicating that the
command was executed correctly.

In cases where there is a delay in interrupt handling, this enhancement
ensures that we still verify whether the hardware has completed the
command. If the command is confirmed to be processed successfully by
checking the CH_ctrl_irq bit, we return success instead of an error.
This improvement enhances the reliability and performance of command
execution.

Change-Id: Ibee70a52245fe1beb9c22ebba950a233fd563ce2
Signed-off-by: Somesh Dey <quic_somedey@quicinc.com>