scsi: ufs: qcom: Fix QoS potential issues
If the "mask" of the QoS DT node is not provided, the driver
will use the uninitialized mask variable which may cause stability
issues. In case the user does not provide proper "mask" QoS DT
entry, the fix is to exit the ufs_qcom_qos_init()
with QoS and cpu frequency scaling features disabled.

Correct the handling of the irq_affinity_support flag in the
ufs_qcom_parse_irq_affinity() function. In case of errors in
parsing of the esi-affinity-mask logic, it may return without
updating the irq_affinity_support flag.

Improve readability. Validating QoS DT prior to allocating
memory for QoS, adding the print of the "mask" in the errors
cases for debug, etc.

Change-Id: I28774dc667c2c7323d501785fcd6f60910bbf003
Signed-off-by: Bao D. Nguyen <quic_nguyenb@quicinc.com>