sched/walt: Update waltgov on first enqueue of window
Currently, task accounting is handled by subtracting task contributions
during dequeue and adding them during enqueue. This method was designed
to avoid holding two runqueue locks for every migration, whether
intercluster or intracluster.

However, this introduces a potential issue where subtraction occurs,
but before enqueue takes place, a window rollover happens. This leads
to a low cpufreq vote because the dequeued task's contributions are
not accounted for at rollover time when determining the vote for
the next window.

The best approach to resolve this without introducing too many cpufreq
updates is to perform a cpufreq update on the first enqueue on each
CPU after a window rollover, provided the predictive load is high
enough as determined by do_pl_notif().

Change-Id: I5d175e179b634979555fed76c77db471f0981c5a
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>