qrtr: genpool: Add lock between setup and reboot
There are instances where the threads for setup worker and reboot cb run
in parallel and both are attempting to update the state machine. This
causes each thread to assume the state machine was updated as expected
leading to the state machine to be out of sync. Add locks between setup
and reboot to prevent the two from updating the state machine
simultaneously.

Change-Id: Ibcde520d757289d4c30bd7675b35fa4a2bc96d5b
Signed-off-by: Tony Truong <quic_truong@quicinc.com>