i2c: i2c-master-msm-geni: update counter values for Fast mode
In I2C fast mode(400KHz), when I2C slave performs clock stretching,
the initial SCL high period was reduced. According to the specifications,
the SCL high period should be at least 600ns, but it was observed to be
580ns in fast mode. To address this, updated the SVE recommended I2C
counter values and with it aligned the SCL high period with the
specifications.

Change-Id: Ie7cc0e17c3ae07a8f6c923417877d4c7688c5a71
Signed-off-by: Anil Veshala Veshala <quic_aveshala@quicinc.com>