mem-buf-gh: add support to use the legacy gunyah interface
When memory is accepted by PVM, that was donated to TVM, not setting of
the GH_RM_MEM_ACCEPT_MAP_IPA_CONTIGUOUS resulting into the clearing of
the GH_RM_MEM_ACCEPT_DONE that can make gh_rm_mem_accept() to fail.

This is reported when using the legacy GH versions where, when sgl
entries are more than 1, the below operations are not supported.
 1) MEM_LEND/SHARE/DONATE.
 2) Multiple MEM_APPEND operations on a handle returned by MEM_LEND.
 3) Accepting via multiple MEM_ACCEPT.

Add support to use the legacy gunyah interface by adding checks for
sglist entries should never be more than 1.

These changes are borrowed from the below commits:
commit 48d703d09aeb ("mem-buf: Return error if scattered memory is
passed to gunyah APIs") #5.10,

commit 0663b7f80117 ("gunyah: Relax physically contiguous memory
reguirements") #5.15 .

Change-Id: I86c0ba7ab9001010b75c9d7520e38ed761d75df4
Suggested-by: Patrick Daly <quic_pdaly@quicinc.com>
Signed-off-by: Charan Teja Kalla <quic_charante@quicinc.com>