sched/walt: Ensure heavy_lock is acquired before swapping
To prevent crashes caused by improperly acquired locks when a pipeline
task dies during execution, it is crucial to always acquire the
heavy_lock before swapping tasks.

Although the heavy_lock was correctly acquired for swapping under the
nr=3 scenario, it was not being acquired under the nr=2 scenario.

Therefore, this update ensures that the heavy_lock is consistently
taken for all task swapping scenarios.

Change-Id: Ie282e12ffe8be4cfb985f04b39c259bf8a1cd629
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>