sched/walt: Fix dead pipeline task accounting
When a pipeline task terminates, its entry must be removed from the
pipeline array, whether it is in heavy_wts or pipeline_wts. It is
crucial to maintain the integrity of the array by ensuring it remains
properly sorted. This means that the array should not contain any gaps,
and entries without tasks should be positioned towards the end of the
array.

To achieve this, the removal process of the dead task should be followed
by a left shift of all remaining elements in the array. This ensures
that the array remains contiguous and that all active tasks are
correctly positioned. Doing so prevents any potential instability or
inefficiency that could arise from having gaps or improperly sorted
entries in the array.

Change-Id: Ia9432212f5809db6d124a07dc24dfd4dfd9394b1
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>