i2c: i2c-msm-geni: Add support for I2C lines recovery
For I2C multi message transfers, the stretch bit is set for each message
to prevent generating an I2C stop in the middle of the transfer, instead
this bit is cleared in the last message of the transfer.

Read and write transfers return an error code due to Memory allocation
and DMA mapping failures. These failures cause the clock line to be low
and subsequent failures in I2C multi message transfers.

To address this, recovery is performed to reset the I2C
lines by performing a I2C stop on bus.

I2C message write transfers with zero length in GSI mode are handled with
immediate dma operation and submitted to the GSI hardware, leading to
timeout errors. GSI can't handle the zero length transfers and so should
not submit zero length write transfers to GSI.
Added logic to check for zero-length I2C message and return from the
transfer if any such messaage is encountered.

For shared use cases, ensure the configuration TRE is invoked for every
I2C message by setting cfg_sent flag to zero.

Change-Id: Iba5afbf15b446847981e6f72902d69a9c3a8929f
Signed-off-by: Jyothi Kumar Seerapu <quic_jseerapu@quicinc.com>