sched/walt: Skip newidle balance if a wakeup in pending on a cpu
When doing newidle balance, we don't account for any pending wakeup on
the corresponding cpu. It might happen that there is a pending wakeup on
that cpu and during our newidle balancing we pull a task from another
cpu. Thus, we end up having two tasks competing for runtime on same cpu.
To avoid this, check if there is a pending wakeup on the rq. If there
is, skip the newidle balance altogether.

Change-Id: I92202716d29f8e247301ba41e5be59420ea8faaa
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>