usb: phy: snps: Update minimum voltage for 1.8V LDO
Currently, the minimum voltage of the 1.8V LDO is below the operational
threshold required for the PHY to function correctly. The regulator
framework aggregates votes from multiple clients, selecting the maximum
among the minimums. However, in the absence of other clients for the
1.8V LDO, the provided minimum voltage becomes the aggregate, which can
lead to instability.

Fix this by updating the minimum voltage to the correct operational
value, ensuring stable performance of the PHY.

Change-Id: I95a8e85b4a3940627d7f95c8aaee36f3288e8a5c
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>