pci: msm: Fix the initialization of GPIOs from DTB
When the driver registers GPIOs based on DTB, the loop currently
iterates only over the number of GPIOs defined in the DTB which picks
GPIOs from predefined array 'pci_dev->gpio[]' in sequence.
If the Index of the GPIO (defined in the DTB) in this predefined
array is beyond the number of GPIOs in DTB, that GPIO would be left
unregistered.

Fix the loop to iterate to all possible GPIOs and initialize.

Correct default non-initialized value for the GPIO number from 0 to
EINVAL, since 0 is a valid GPIO number.

Change-Id: Iaaddc80afab987ee04dd4e998501a5b8a338b519
Signed-off-by: Arnav Kumar Dawn <quic_arnavkum@quicinc.com>