sched/cpufreq_walt: Clean up max fields in waltgov structs
Currently, there is a variable within the waltgov structs to
store the max capacity of the CPU and policy.
Remove this variable to reduce overhead and opt to obtain it on the fly
instead.
Since max capacity of all CPUs in a policy will be the same, remove the
max factor while determining driving CPU and let it be decided solely by
CPU util.

Change-Id: Ia9e9eccfbcca0a1a2f025738aab3dbcb9cce6c10
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>