Makefile: Allow checking CONFIG tokens in DTs
Provide an option to select DT properties based on a kernel config token.
This is based on discussions with Google.

To use this in a DT file, add following line near top of the DT

	include <linux/kconfig.h>

and wrap the property like:

	#if IS_ENABLED(CONFIG_FOO)
	qcom,property = "bar"
	#else
	qcom,property = "baz"
	#endif

Note that this commit must be merged before using the feature in DT.

Change-Id: I12108c35ae800d99f865e788623f2f022ecbad59
Signed-off-by: Sukadev Bhattiprolu <quic_sukadev@quicinc.com>
Signed-off-by: Bibek Kumar Patro <quic_bibekkum@quicinc.com>