sched: walt: ensure heavy tasks are marked as PIPELINE
There is a possibility that the low_latency flag for auto pipeline
doesn't get set if the 'heavy_wts_to_drop' list is empty.
This flag is also used to clean up the 'heavy_wts' list when a task
gets killed. Fix this by ensuring the low_latency flag is set before
checking the 'heavy_wts_to_drop' list.

Change-Id: I6fac18570e7526c389b02da442f8aa01588c077f
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>