dmaengine: gpi: Fix infinite loop and channel handling in GPI_Config
A logical error in gpi_config() occurs when a start channel fails,
causing the error handling loop to incorrectly index gpii_chan.
This can lead to an infinite loop and incorrect command execution.

To resolve this, the loop is adjusted to iterate in reverse from
the last successfully started channel. This ensures that stop and
reset commands are accurately directed to the appropriate channels,
preventing potential infinite loops and ensuring proper error
handling when start channels fails.

Upstream Reference :
https://lore.kernel.org/all/20210103135738.3741123-1-arnd@kernel.org/.

Change-Id: If0ed9fc8246bdb26cf8d1eff5d0887aac9886232
Signed-off-by: Prasanna S <quic_prass@quicinc.com>