slimbus: qcom-ngd-ctrl: Skip runtime resume if early service down is initiated
If the PDR service early down is initiated and runtime resume is invoked
before the actual service down, it causes a delay in the response of the
ADSP power-on request over the QMI interface. This further leads to a delay
in returning the PDR callback function.

To avoid this, do not queue new power-on requests if the service early
shutdown is triggered but the actual shutdown is still pending.

Change-Id: I85aaf51e5c8d7c45b6a0d73b02f06c5cc1732290
Signed-off-by: Mehul Raninga <quic_mraninga@quicinc.com>